+++ varinfo/generate_umm_var.py
@@ -18,13 +18,14 @@ import re
 
 from cmr import CMR_UAT
 
-from varinfo import VarInfoFromNetCDF4
+from varinfo import VarInfoFromNetCDF4, VarInfoFromDmr
 from varinfo.cmr_search import (
     CmrEnvType,
     download_granule,
     get_granule_link,
     get_granules,
     get_edl_token_header,
+    get_dmr_xml_url,
 )
 from varinfo.umm_var import get_all_umm_var, publish_all_umm_var
 
@@ -40,6 +41,7 @@ def generate_collection_umm_var(
     cmr_env: CmrEnvType = CMR_UAT,
     publish: bool = False,
     config_file: str | None = None,
+    use_dmr: bool = False,
 ) -> UmmVarReturnType:
     """Run all the of the functions for downloading and publishing
     a UMM-Var entry to CMR given:
@@ -64,8 +66,13 @@ def generate_collection_umm_var(
         collection_concept_id, cmr_env=cmr_env, auth_header=auth_header_edl_token
     )
 
-    # Get the data download URL for the most recent granule (NetCDF-4 file)
-    granule_link = get_granule_link(granule_response)
+    if use_dmr:
+        # Get OPeNDAP data URL with `.dml.xml` appended
+        granule_link = get_dmr_xml_url(granule_response)
+
+    else:
+        # Get the data download URL for the most recent granule (NetCDF-4 file)
+        granule_link = get_granule_link(granule_response)
 
     with TemporaryDirectory() as temp_dir:
         # Download file to runtime environment
@@ -73,11 +80,15 @@ def generate_collection_umm_var(
             granule_link, auth_header_edl_token, out_directory=temp_dir
         )
 
-        # Parse the granule with VarInfo to map all variables and relations:
-        var_info = VarInfoFromNetCDF4(local_granule, config_file=config_file)
+        if use_dmr:
+            # Parse the granule with VarInfoFromDmr to map all variables and relations:
+            var_info = VarInfoFromDmr(local_granule, config_file=config_file)
+        else:
+            # Parse the granule with VarInfoFromNetCDF4 to map all variables and relations:
+            var_info = VarInfoFromNetCDF4(local_granule, config_file=config_file)
 
-        # Generate all the UMM-Var records:
-        all_umm_var_records = get_all_umm_var(var_info)
+    # Generate all the UMM-Var records:
+    all_umm_var_records = get_all_umm_var(var_info)
 
     if publish:
         # Publish to CMR and construct an output object that is a list of